+++ tests/interchain/integrator/export_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/cosmos/gaia/v28/tests/interchain/chainsuite"
+	"github.com/cosmos/gaia/v29/tests/interchain/chainsuite"
 	"github.com/cosmos/interchaintest/v10"
 	"github.com/cosmos/interchaintest/v10/chain/cosmos"
 	"github.com/cosmos/interchaintest/v10/ibc"
@@ -13,8 +13,7 @@ import (
 )
 
 const (
-	maxValidators          = 5
-	maxConsensusValidators = 4
+	maxValidators = 5
 )
 
 type ExportSuite struct {
@@ -50,6 +49,7 @@ func (s *ExportSuite) TestExportAndImportValidators() {
 	zero := 0
 	newSpec := &interchaintest.ChainSpec{
 		Name:          "gaia",
+		ChainName:     "gaia-reimport",
 		Version:       s.Env.NewGaiaImageVersion,
 		NumValidators: &chainsuite.SixValidators,
 		NumFullNodes:  &zero,
@@ -65,15 +65,21 @@ func (s *ExportSuite) TestExportAndImportValidators() {
 		s.Require().NoError(err)
 		validators = append(validators, validator)
 	}
-	for i := 0; i < maxValidators; i++ {
+
+	// Check that the correct number of validators are bonded and that the next validator is unbonded (i.e. not in the active set).
+	maxValidatorsJSON, err := newChain.QueryJSON(s.GetContext(), "params.max_validators", "staking", "params")
+	s.Require().NoError(err)
+	maxValidatorsParam := int(maxValidatorsJSON.Int())
+
+	for i := range maxValidatorsParam {
 		s.Require().Equal(stakingtypes.Bonded, validators[i].Status)
 	}
-	s.Require().Equal(stakingtypes.Unbonded, validators[maxValidators].Status)
+	s.Require().NotEqual(stakingtypes.Bonded, validators[maxValidatorsParam].Status)
 
 	vals, err := newChain.QueryJSON(s.GetContext(), "validators", "tendermint-validator-set")
 	s.Require().NoError(err)
-	s.Require().Equal(maxConsensusValidators, len(vals.Array()), vals)
-	for i := 0; i < maxConsensusValidators; i++ {
+	s.Require().Equal(maxValidatorsParam, len(vals.Array()), vals)
+	for i := range maxValidatorsParam {
 		valCons := vals.Array()[i].Get("address").String()
 		s.Require().NoError(err)
 		s.Require().Equal(s.Chain.ValidatorWallets[i].ValConsAddress, valCons)
@@ -84,7 +90,6 @@ func TestExport(t *testing.T) {
 	genesis := chainsuite.DefaultGenesis()
 	genesis = append(genesis,
 		cosmos.NewGenesisKV("app_state.staking.params.max_validators", maxValidators),
-		cosmos.NewGenesisKV("app_state.provider.params.max_provider_consensus_validators", maxConsensusValidators),
 	)
 	s := &ExportSuite{
 		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{